Add files via upload
So, I added the pause button on the screen , by drawing it out with basic shapes (1 circle and 2 rectangles). You can find on #6.6 Draw pause button
Also , I created a pause function which is basically a while loop that waits the player to press the button again or press Esc to unpause.
I didn't add a pause menu yet , there is just a gray screen with the title Paused.
I also made the pause function to return the amount of time that it was paused , so we can add back to the countdown timer , so it doesn't reach the end while staying paused.